pyproject.toml
[tool.poetry]
name = "elune"
version = "0.1.0"
description = ""
authors = ["Apor Bokor <aporbokor@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cmd2 = "^2.4.3"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
elune = "elune.__main__:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
